+++ cds/partial_sums.h
@@ -5,23 +5,23 @@
 
 namespace cds {
     class partial_sums : public fixed_length_vector {
-        std::vector<unsigned int> samples;
+        std::vector<uint64_t> samples;
 
     public:
-        unsigned int rate;
+        uint64_t rate;
 
         partial_sums(
-            unsigned int length,
-            unsigned int rate = 4,
-            unsigned int size = 0,
+            uint64_t length,
+            uint64_t rate = 4,
+            uint64_t size = 0,
             bool is_rapid = false
         );
 
-        void resize(unsigned int size);
-        unsigned int vector_size();
-        void write(unsigned int index, unsigned int value);
+        void resize(uint64_t size);
+        uint64_t vector_size();
+        void write(uint64_t index, uint64_t value);
 
-        unsigned int sum(unsigned int index);
-        unsigned int search(unsigned int value);
+        uint64_t sum(uint64_t index);
+        uint64_t search(uint64_t value);
     };
 }
