+++ cds/variable_length_vector/base.h
@@ -4,12 +4,12 @@
 namespace cds {
     class variable_length_vector {
     public:
-        unsigned int size;
+        uint64_t size;
 
         virtual ~variable_length_vector() {}
 
-        virtual unsigned int vector_size() = 0;
-        virtual unsigned int read(unsigned int index) = 0;
-        virtual void push_back(unsigned int value) = 0;
+        virtual uint64_t vector_size() = 0;
+        virtual uint64_t read(uint64_t index) = 0;
+        virtual void push_back(uint64_t value) = 0;
     };
 }
