+++ cds/variable_length_vector/dense_pointers.h
@@ -9,15 +9,15 @@ namespace cds {
     class dense_pointers : virtual public variable_length_vector {
         bit_vector bv;
         fixed_length_vector offsets;
-        std::vector<unsigned int> samples;
+        std::vector<uint64_t> samples;
 
     public:
-        unsigned int rate;
+        uint64_t rate;
 
-        dense_pointers(unsigned int max_length, unsigned int rate = 4);
+        dense_pointers(uint64_t max_length, uint64_t rate = 4);
 
-        unsigned int vector_size();
-        unsigned int read(unsigned int index);
-        void push_back(unsigned int value);
+        uint64_t vector_size();
+        uint64_t read(uint64_t index);
+        void push_back(uint64_t value);
     };
 }
