+++ cds/variable_length_vector/direct_access_codes.cpp
@@ -2,15 +2,15 @@
 
 
 namespace cds {
-    direct_access_codes::direct_access_codes(unsigned int length) {
+    direct_access_codes::direct_access_codes(uint64_t length) {
         this->length = length;
         this->size   = 0;
         this->markers.push_back(partial_sums(1));
         this->blocks.push_back(fixed_length_vector(this->length));
     }
 
-    unsigned int direct_access_codes::vector_size() {
-        unsigned int vector_size = 0;
+    uint64_t direct_access_codes::vector_size() {
+        uint64_t vector_size = 0;
         for (auto marker : this->markers) {
             vector_size += marker.vector_size();
         }
@@ -20,9 +20,9 @@ namespace cds {
         return vector_size;
     }
 
-    unsigned int direct_access_codes::read(unsigned int index) {
-        unsigned int value = 0;
-        unsigned int depth = 0;
+    uint64_t direct_access_codes::read(uint64_t index) {
+        uint64_t value = 0;
+        uint64_t depth = 0;
         while (1) {
             value |= (this->blocks[depth].read(index) << (depth * this->length));
             if (this->markers[depth].read(index) == 0) {
@@ -33,14 +33,14 @@ namespace cds {
         }
     }
 
-    void direct_access_codes::push_back(unsigned int value) {
-        unsigned int mask = 0;
-        for (unsigned int i = 0; i < this->length; i++) {
+    void direct_access_codes::push_back(uint64_t value) {
+        uint64_t mask = 0;
+        for (uint64_t i = 0; i < this->length; i++) {
             mask = (mask << 1) | 1;
         }
 
-        unsigned int depth = 0;
-        unsigned int index = this->size;
+        uint64_t depth = 0;
+        uint64_t index = this->size;
         while (1) {
             if (this->markers.size() == depth) {
                 this->markers.push_back(partial_sums(1));
