+++ cds/variable_length_vector/sampled_pointers.cpp
@@ -2,26 +2,26 @@
 
 
 namespace cds {
-    sampled_pointers::sampled_pointers(unsigned int rate) {
+    sampled_pointers::sampled_pointers(uint64_t rate) {
         this->rate = rate;
         this->size = 0;
     }
 
-    unsigned int sampled_pointers::vector_size() {
+    uint64_t sampled_pointers::vector_size() {
         return this->bv.vector_size() + this->samples.size();
     }
 
-    unsigned int sampled_pointers::read(unsigned int index) {
-        unsigned int begin = this->samples[index / this->rate];
-        unsigned int values_to_read = index % this->rate;
+    uint64_t sampled_pointers::read(uint64_t index) {
+        uint64_t begin = this->samples[index / this->rate];
+        uint64_t values_to_read = index % this->rate;
 
         while (true) {
-            unsigned int zeros_length = 0;
+            uint64_t zeros_length = 0;
             while (this->bv.bit_read(begin) == 0) {
                 zeros_length++;
                 begin++;
             }
-            unsigned int end = begin + zeros_length + 1;
+            uint64_t end = begin + zeros_length + 1;
 
             if (values_to_read == 0) {
                 return this->bv.bits_read(begin, end, true) - 1;
@@ -31,17 +31,17 @@ namespace cds {
         }
     }
 
-    void sampled_pointers::push_back(unsigned int value) {
+    void sampled_pointers::push_back(uint64_t value) {
         if (this->size % this->rate == 0) {
             this->samples.push_back(this->bv.size);
         }
         this->size++;
 
         value++;
-        unsigned int value_length = bits_length(value);
+        uint64_t value_length = bits_length(value);
 
-        unsigned int begin = this->bv.size + value_length - 1;
-        unsigned int end = begin + value_length;
+        uint64_t begin = this->bv.size + value_length - 1;
+        uint64_t end = begin + value_length;
 
         this->bv.resize(end);
         this->bv.bits_write(begin, end, value, true);
